requirements.txt
PySide6>=6.4.0
psutil>=5.9.0
Pillow>=9.0.0
pyinstaller>=5.0.0
